back/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-lazy-fixture = "^0.6.3"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
src_paths = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]